requirements.txt
Flask
Flask-Login
Flask-Session
msal
python-dotenv
